+++ src/tokdash/__init__.py
@@ -1,3 +1,3 @@
 """Tokdash package."""
 
-__version__ = "1.1.2"
+__version__ = "1.1.3"